requirements.txt
click==8.1.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
zipp==3.8.0
Flask_SQLAlchemy==2.5.1
Flask==2.0.1
SQLAlchemy==1.4.22
ffmpeg-python==0.2.0
